requirements.txt
redis
celery
django-admin
djangorestframework